+++ R/globals.R
@@ -1,6 +1,8 @@
 utils::globalVariables(c("COMID", "HTML", "ID", "acres", "characteristic_id",
                          "characteristic_value", "chars", "data", "featureid",
                          "group_by", "mi", "req", "sessionEnded", "setProgress",
-                         "tags", "withProgress",".", "rowid")
+                         "tags", "withProgress",".", "rowid", "site_no", "datetime",
+                         "flow_chr", "StatisticsStatusColorFill", "StatisticsStatusColorStroke",
+                         "StatisticsStatusDescription")
 )