+++ R/streamnetwork.R
@@ -37,13 +37,6 @@
 #' nhdplus_data <- get_stream_network_interactively()
 #' }
 #'
-#' @importFrom shiny NS tagList
-#' @importFrom dplyr filter select mutate slice_max ungroup rename group_by
-#' @importFrom sf st_area st_transform st_geometry st_as_sf
-#' @importFrom scales comma
-#' @importFrom leaflet addPolygons addPolylines addCircles
-#' @importFrom shiny observeEvent reactive actionButton HTML req withProgress setProgress tags
-#' @importFrom dplyr "%>%"
 #'
 #'
 get_stream_network_interactively <- function(ns = 'basin-ui',