+++ R/usgsinst.R
@@ -1,8 +1,8 @@
-#' Get Watershed Basin Interactively
+#' Get United States Geologic Survey (USGS) Instantaneous Flow Values Interactively
 #'
-#' @description This function allows the user to delineate watershed basins interactively with a
-#' shiny app. It uses the {elevatr} package to acquire the Digital Elevation Model (DEM)
-#' and {whitebox} package to delineate the basin (see details).
+#' @description This function allows the user to select United States Geologic Survey (USGS) stations
+#' and get back instantaneous flow values based on control `number of days from now`. It uses the USGS
+#' Water Services to get the values as well as the USGS Dashboard to get current conditions (circle markers on map).
 #' @param ns \code{string} name for the Shiny \code{namespace} to use.  The \code{ns}
 #'          is unlikely to require a change.
 #' @param viewer \code{function} for the viewer.  See Shiny \code{\link[shiny]{viewer}}.
@@ -12,52 +12,37 @@
 #'          Firefox is an exception. See Details for instructions on how to enable this
 #'          behaviour in Firefox.
 #' @param title \code{string} to customize the title of the UI window.  The default
-#'          is "Delineate Basin".
+#'          is "Get USGS Instantaneous Flow Values".
 #' @param ... other arguments to \code{leafletOutput()} in module.
-#' @return A list of sf objects that contain watershed polygons
-#'         the user collected during shiny session.
+#' @return A data.frame that contains flow values based on the station(s) selected
+#'         during shiny session.
 #'
-#' @note The marker will only work for the most current stream raster. You can have multiple areas but
-#' you need to make sure that you are on the most current raster or the app will crash.
+#' @note You can select multiple stations but the `number of days from now` control
+#' will take the final number when you select the `done` button. The information from the `hover` details is not included in
+#' the data.frame that is returned, e.g. rate of change, percentile description.
 #'
 #' @details
-#' **This function will throw an error if you don't draw the bounding box (rectangle) first.**
-#' Once the user has drawn the bounding box then you can use the marker as a pour point location.
 #'
 #' **Steps**
 #'
-#' 1. Input a well-suited DEM zoom level, stream threshold (cells) and snapping distance.
-#' 2. Draw bounding box (rectangle).
-#' 3. Use marker to place pour point(s).
-#' 4. If necessary, change 'Cell Threshold' to change drainage density.
-#' 5. Repeat steps 1-4 if needed.
-#' 6. When finished, press 'done' and basins will be saved as a list in local environment.
+#' 1. Select the sites you want to retrieve.
+#' 2. Make sure you have the right days.
+#' 3. When finished, press 'done' and sites instantaneous flow values will be saved to
+#' a data.frame in local environment.
 #'
-#' In addition, this function uses both `whitebox::wbt_feature_preserving_smoothing()` and `whitebox::wbt_breach_depressions()`
-#' prior to running the flow direction and flow accumulation (both d8) algorithms.
 #' @export
 #' @examples
 #'
 #' if(interactive()){
-#' nhdplus_data <- get_basin_interactively()
+#' nhdplus_data <- get_usgs_instantaneous_interactively()
 #' }
 #'
-#' @importFrom shiny NS tagList
-#' @importFrom dplyr filter select mutate slice_max ungroup rename group_by
-#' @importFrom sf st_area st_transform st_geometry st_as_sf
-#' @importFrom scales comma
-#' @importFrom leaflet addPolygons addPolylines addCircles
-#' @importFrom shiny observeEvent reactive actionButton HTML req withProgress setProgress tags
-#' @importFrom dplyr "%>%"
 #'
 #'
-get_usgs_instantaneous_interactively <- function(ns = 'basin-ui',
+get_usgs_instantaneous_interactively <- function(ns = 'usgsiv-ui',
                                     viewer = shiny::paneViewer(),
-                                    title = 'Delineate Basin',
+                                    title = 'Get USGS Instantaneous Flow Values',
                                     ...) {
-
-  #spherical geometry switched off
-  #sf::sf_use_s2(FALSE)
 
   ## Some code hijacked from mapedit throughout; to get miniUI look, etc
 
@@ -122,9 +107,32 @@ $(document).on('shiny:disconnected', function() {
     #used to stop the app via button and retain selections
     observeEvent(input$done, {
 
-      shiny::stopApp(
-        values$basin_data_list
+      p <- shiny::Progress$new()
+      p$set(message = "Downloading instantaneous values...",
+            detail = "This may take a little bit...",
+            value = 1/2)
+
+      promises::future_promise({
+
+        sites <- values$df[as.logical(values$df$selected),1]
+        period <- values$period
+        list(sites = sites, period = period)
+
+        }) %...>% {
+
+
+        suppressWarnings(values$final_dv <- get_iv(.[['sites']], .[['period']]))
+
+
+
+        shiny::stopApp(
+
+        values$final_dv
+
       )
+
+        }  %>%
+        finally(~p$close())
 
     })