+++ NEWS.md
@@ -1,7 +1,7 @@
 # gwavr 0.2.0
 
 * Added a new function `get_basin_interactively()` and new modules `basinModUI` and `basinMod`.
-* Added three new dependencies: terra, elevatr, and whitebox.
+* Added three new dependencies: 'terra', 'elevatr', and 'whitebox'.
 
 # gwavr 0.1.0